modify basic data structures
